add tricks to avoid setting type params explicitly
suddenly I came up with the idea to implement a function only for
specialized structs and use it to help rustc to deduce the type.
By calling the dummy function that is implemented only for the
specialized structs, the type parameter can be determined.
Default type parameter didn't work for this purpose, but it works well.